feat(api): add OpenAPI documentation to entities
- Add Swagger/OpenAPI annotations to domain entities
  - Product and related DTOs
  - User and related DTOs
  - Request/Response models

- Add detailed field documentation
  - Add descriptive comments for each field
  - Include meaningful example values
  - Document nullable fields in update DTOs

- Improve type consistency
  - Use rust_decimal::Decimal for monetary values
  - Add proper schema examples for all fields
  - Ensure consistent type usage across DTOs

Technical details:
- Add ToSchema derive macro for OpenAPI generation
- Add field-level documentation with /// comments
- Add schema examples with #[schema(example = '..')] attributes

+++ src/interfaces/http/requests/product_requests.rs
@@ -1,17 +1,35 @@
-use serde::Deserialize;
+use rust_decimal::Decimal;
+use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
 
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, ToSchema)]
 pub struct CreateProductRequest {
+    /// The name of the product
+    #[schema(example = "iPhone 14 Pro")]
     pub name: String,
+    /// Detailed description of the product
+    #[schema(example = "Latest iPhone model with dynamic island")]
     pub description: String,
-    pub price: f64,
+    /// The price of the product
+    #[schema(example = "999.99")]
+    pub price: Decimal,  // Change from f64 to Decimal
+    /// Initial stock quantity
+    #[schema(example = "100")]
     pub stock: i32,
 }
 
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, ToSchema)]
 pub struct UpdateProductRequest {
+    /// Updated name of the product
+    #[schema(example = "iPhone 14 Pro Max")]
     pub name: Option<String>,
+    /// Updated description of the product
+    #[schema(example = "Updated description for iPhone")]
     pub description: Option<String>,
-    pub price: Option<f64>,
+    /// Updated price of the product
+    #[schema(example = "1099.99")]
+    pub price: Option<Decimal>,  // Change from Option<f64> to Option<Decimal>
+    /// Updated stock quantity
+    #[schema(example = "50")]
     pub stock: Option<i32>,
 }
